requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
requests==2.32.3
python-dotenv==1.0.1
openai==1.57.4
langchain==0.3.11
langchain-community==0.3.11
langchain-openai==0.2.12
tiktoken==0.8.0
faiss-cpu==1.9.0.post1

# webserver
django==5.1.4
django-cors-headers==4.6.0

# telegram
python-telegram-bot